ollama/requirements.txt
langchain_ollama==0.3.1
python-dotenv==1.1.0
